requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-advanced-filters==1.3.0
django-braces==1.14.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-formtools==2.3
django-import-export==2.5.0
django-redis==5.0.0
django-reversion==4.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.9.2
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
prompt-toolkit==3.0.19
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==1.7.1
PyMySQL==1.0.2
python-alipay-sdk==1.8.0
pytz==2021.1
PyYAML==5.4.1
qcloudsms-py==0.1.4
redis==3.5.3
requests==2.26.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
